etls/requirements.txt
dbt-core==1.3.1
dbt-postgres
dbt-bigquery==1.3.0
google-cloud-bigquery